SwapMoveSelector: fix bug: return A-B, but not B-A (non random)

+++ drools-planner-core/src/main/java/org/drools/planner/core/heuristic/selector/move/composite/CartesianProductMoveSelector.java
@@ -64,7 +64,7 @@ public class CartesianProductMoveSelector extends CompositeMoveSelector {
     public long getSize() {
         long size = 1L;
         for (MoveSelector moveSelector : childMoveSelectorList) {
-            size *= moveSelector.getSize();
+            size *= (long) moveSelector.getSize();
         }
         return size;
     }
